// src/app/api/generated/models/customers-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CustomersDto {
  address?: (string | null);
  city?: (string | null);
  companyName: (string | null);
  contactName?: (string | null);
  contactTitle?: (string | null);
  country?: (string | null);
  customerID: (string | null);
  fax?: (string | null);
  phone?: (string | null);
  postalCode?: (string | null);
  region?: (string | null);
}